samples: split resource file into individual files
